ajout interfaces axios

+++ cefiishop/src/main/java/com/cefiishop/model/UserBehavior.java
@@ -12,22 +12,22 @@ public class UserBehavior {
 
     @Id
     private String id;
-    private Long userId;
+    private Integer userId;
     private List<ProductView> viewedProducts = new ArrayList<>();
 
     public String getId() { return id; }
     public void setId(String id) { this.id = id; }
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
+    public Integer getUserId() { return userId; }
+    public void setUserId(Integer userId) { this.userId = userId; }
     public List<ProductView> getViewedProducts() { return viewedProducts; }
     public void setViewedProducts(List<ProductView> viewedProducts) { this.viewedProducts = viewedProducts; }
 
     public static class ProductView {
-        private Long productId;
+        private Integer productId;
         private LocalDateTime viewedAt = LocalDateTime.now();
 
-        public Long getProductId() { return productId; }
-        public void setProductId(Long productId) { this.productId = productId; }
+        public Integer getProductId() { return productId; }
+        public void setProductId(Integer productId) { this.productId = productId; }
         public LocalDateTime getViewedAt() { return viewedAt; }
         public void setViewedAt(LocalDateTime viewedAt) { this.viewedAt = viewedAt; }
     }
